+++ poem-openapi/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# Unreleased
+
+- Add support for custom validator.
+
 # [1.2.29] 2022-1-11
 
 - Add `NewType` macro. [#159](https://github.com/poem-web/poem/issues/159)